requirements.txt
networkx>=2.5
matplotlib>=3.3.0
pyyaml>=5.4.0